feat: add useSudo option
