+++ lib/gpg.ts
@@ -12,77 +12,18 @@ import { homedir } from "os";
 import path from "path";
 import { v4 as uuid } from "uuid";
 import { Stream } from "stream";
-import { IStreamingOptions, spawnGPG, streaming } from "./spawnGPG";
+import { spawnGPG, streaming } from "./spawnGPG";
+import { IGpgKey, IGpgOptions, IStreamingOptions } from "./types";
+import { parseKeysFromOutput } from "./parsers";
+
 const keyRegex = /^gpg: key (.*?):/;
-// eslint-disable-next-line no-control-regex
-const emailRegex = /(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])/gi;
-
-type SpawnGpgFn = (input: string, args: string[]) => Promise<void | Buffer>;
-type StreamingFn = (
-  options: IStreamingOptions,
-  args: string[]
-) => Promise<fs.WriteStream>;
-interface IGpgKey {
-  created_at: string | Date;
-  expires_at: string | Date;
-  id: string;
-  username: string;
-  keygrip: string;
-}
-
-const parseKeysFromOutput = (output: string): IGpgKey[] => {
-  const [first, second] = output.split(/-----+/).map((l) => l.trim());
-  const ourFocus = second || first;
-  const lines = ["", ...ourFocus.split("\n").map((l) => l.trim())];
-  return lines.reduce((keys, line) => {
-    if (line === "") {
-      return keys.concat({});
-    }
-    const lastKey = keys[keys.length - 1] as Partial<IGpgKey>;
-    if (line.startsWith("pub")) {
-      const [created_at, expires_at] = line
-        .match(/(\d{4}-\d{2}-\d{2}) .+ (\d{4}-\d{2}-\d{2})/)
-        .slice(1);
-      lastKey.created_at = created_at;
-      lastKey.expires_at = expires_at;
-    } else if (line.startsWith("uid")) {
-      const [ourFocus] = line.split("]").slice(1);
-      const [name] = ourFocus.match(/(.+ <(.+)>)/)?.slice(1) || [];
-      const [email] = ourFocus.match(emailRegex);
-      lastKey.username = name?.trim().replace('"', "") || email;
-    } else if (line.trim().startsWith("Keygrip")) {
-      const [keygrip] = line.split(" = ").slice(1);
-      lastKey.keygrip = keygrip;
-    } else if (!line.startsWith("sub")) {
-      lastKey.id = line;
-    }
-    return keys;
-  }, []);
-};
 
 export const GPG_UNIX_BASE_DIR = `${homedir()}/.gnupg`;
 export const GPG_WINDOWS_BASE_DIR = `${homedir()}\\AppData\\Roaming\\gnupg`;
 
 export class GpgService {
   constructor(
-    private options: {
-      spawnGPG?: SpawnGpgFn;
-      streaming?: StreamingFn;
-      basedir?: string;
-      tempFolderPath?: string;
-      reader?: {
-        readFile: (filePath: string) => Promise<Buffer>;
-        readFileString: (filePath: string) => Promise<string>;
-      };
-      writer?: {
-        writeFile: (
-          filePath: string,
-          content: string | Buffer
-        ) => Promise<void>;
-        unlink: (filePath: string) => Promise<void>;
-      };
-      idFactoryFn?: () => string;
-    }
+    private options: IGpgOptions
   ) {}
 
   /**
@@ -93,7 +34,7 @@ export class GpgService {
       "--homedir",
       this.options.basedir,
       ...args,
-    ]) as Promise<Buffer>;
+    ], this.options) as Promise<Buffer>;
   }
 
   /**
@@ -109,7 +50,7 @@ export class GpgService {
       "--homedir",
       this.options.basedir,
       ...args,
-    ]);
+    ], this.options);
   }
 
   setTempFolderPath(tempFolderPath: string): GpgService {
@@ -119,6 +60,11 @@ export class GpgService {
 
   setBaseDir(basedir: string): GpgService {
     this.options.basedir = basedir;
+    return this;
+  }
+
+  useSudo(value: boolean): GpgService {
+    this.options.useSudo = value;
     return this;
   }
 
@@ -498,5 +444,6 @@ export const gpg = new GpgService({
   tempFolderPath: "./",
   idFactoryFn: uuid,
   basedir: GPG_UNIX_BASE_DIR,
+  useSudo: false
 });
 export default gpg;
