+++ lib/index.ts
@@ -2,5 +2,6 @@ import gpg from "./gpg";
 
 export * from "./spawnGPG";
 export * from "./gpg";
+export * from "./types";
 
 export default gpg;
